requirements.txt
fastapi
uvicorn
requests
openai
Pillow
